Moved IQRSearch webapp blueprint from __init__ file into its own module
